+++ src/zeus2_toolbox/pipeline.py
@@ -328,6 +328,8 @@ def transmission_pixel(freq, pwv, elev=60, r=1000, d_freq=0.8):
 
 def get_transmission_raw_obs_array(array_map, pwv, elev=60):
     """
+    CAUTION: currently not working
+    TODO: modify to be compatible with the new ArrayMap.init_wl()
     get an ObsArray object corresponding to the raw sky transmission of the
     array, the array_map_ of the output obe_array object will correspond to the
     frequency of transmission_raw_range() output instead of input array_map.